fix(security): address critical security and infrastructure issues from audit
- Remove hardcoded dev-only secret defaults from fiscal-backend and beeminipos-backend config (BLE_SIGNING_KEY, INTEGRATION_SECRET_PEPPER, INTEGRATION_ENCRYPTION_KEY_BASE64, AUTH_HMAC_KEY, WEBHOOK_VERIFICATION_KEY)
- Add sslmode=disable validation: both backends now refuse to start in prod with unencrypted DB connections
- Add sslmode=require DATABASE_URL overrides in both prod compose overlays
- Isolate Device CA private key from API container: move to Docker secret in prod overlay, cert-only volume mount; PKI dir mount moved to dev overlay
- Add restart: unless-stopped to PostgreSQL in both base compose stacks
- Strip all dev-only secret defaults from base compose files; move to dev overlays (EMQX JWT secret, dashboard password, RabbitMQ password, MinIO password, DB passwords, BLE key)
- Fix auth bypass: panic on missing AUTH_HMAC_KEY/OIDC provider at startup
- Fix authorizeSale() tenant isolation: deny when no tenant claim
- Remove non-prod auth bypass in beeminipos-backend handler
- Replace timestamp+counter newID() with crypto/rand UUID to prevent cross-instance collisions
- Add rate limiter entry eviction after window expiry
- Enforce RSA key minimum 2048 bits in OIDC verifier
- Force SMTP over explicit TLS (tls.Dial) in emailauth service
- Fix IoT: Daisy _writeAll() uint8_t offset overflow, Tremol silent data truncation, Datecs unchecked snprintf result, deviceInfo bufLen uint16_t
- Fix miniposweb: IndexedDB singleton pattern, navigator.onLine guard, token refresh respecting expires_in, cart matching by index not name, shift existence check, BLE frame MAX validation

+++ IoT/common-modules/daisy/DaisyProtocol.cpp
@@ -82,11 +82,11 @@ size_t DaisyProtocol::sendPacket(uint8_t cmd, const char* data, uint8_t dataLen)
 
 bool DaisyProtocol::_writeAll(const uint8_t* data, uint8_t len) {
     if (data == nullptr || len == 0) return false;
-    uint8_t offset = 0;
+    uint16_t offset = 0;
     while (offset < len) {
         size_t written = _serial.write(data + offset, len - offset);
         if (written == 0 || written > (size_t)(len - offset)) return false;
-        offset = (uint8_t)(offset + written);
+        offset = (uint16_t)(offset + written);
     }
     return true;
 }
